streamlit_app/requirements.txt
streamlit==1.40.0
python-dotenv==1.0.0
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.36
faiss-cpu==1.7.4
python-docx==1.0.1
pydantic==2.6.3
langchain-text-splitters==0.0.1
sentence-transformers==3.0.0
huggingface-hub==0.22.1
